requirements.txt
# Sigyn AI - Requirements

# Core dependencies
ultralytics==8.4.14
torch>=2.0.0
torchvision>=0.15.0
roboflow>=1.1.0
onnx>=1.14.0
onnxsim>=0.4.0
onnxruntime>=1.15.0
pyyaml>=6.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Optional: OAK-D support
depthai>=2.22.0.0
blobconverter>=1.4.0

# Optional: Development tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

# Optional: TensorBoard
tensorboard>=2.13.0

# Note: For TensorRT (Jetson), install on device:
#   pip install tensorrt  # Via NVIDIA JetPack
